Add explicit children
The implicit children property on components are deprecated and are
being removed in React 18. Some type libs might adapt to this change
prematurely as well, possibly breaking dependents
(@types/styled-components is an example of this).

Add explicit children properties to those components that need it now to
avoid this and prepare for future support of React 18.

See https://solverfox.dev/writing/no-implicit-children/

Signed-off-by: Niklas Holm <niklas.z.holm@telia.no>